% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissHistMunData-package.r
\docType{data}
\name{canton}
\alias{canton}
\title{Canton data}
\format{An object of class \code{data.frame} with 26 rows and 4 columns.}
\usage{
canton
}
\description{
Canton data
}
\keyword{datasets}

