% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissHistMunData-package.r
\docType{data}
\name{district}
\alias{district}
\title{District data}
\format{An object of class \code{data.frame} with 296 rows and 13 columns.}
\usage{
district
}
\description{
District data
}
\keyword{datasets}

