% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissHistMunData-package.r
\docType{data}
\name{municipality}
\alias{municipality}
\title{Municipality data}
\format{An object of class \code{data.frame} with 5615 rows and 15 columns.}
\usage{
municipality
}
\description{
Municipality data
}
\keyword{datasets}

