% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwissHistMunData-package.r
\docType{package}
\name{SwissHistMunData-package}
\alias{SwissHistMunData}
\alias{SwissHistMunData-package}
\title{Historicized Swiss Municipalities}
\description{
Contains historicized municipality data for Switzerland from 1960
onwards.
}
\examples{
cantons
district_mutations
municipality_mutations
}

